+++ src/components/ResearchProblemsBox/ProblemsDropdownFilter.js
@@ -1,81 +1,68 @@
 import { useState } from 'react';
-import { Button, DropdownMenu, DropdownItem, FormGroup, Label, Input, UncontrolledButtonDropdown, DropdownToggle } from 'reactstrap';
+import { Button, FormGroup, Label, Input } from 'reactstrap';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons';
 import Tippy from '@tippyjs/react';
 import { stringifySort } from 'utils';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 
 const ProblemsDropdownFilter = ({ sort, isLoading, includeSubFields, setSort, setIncludeSubFields }) => {
     const [tippy, setTippy] = useState({});
+    const isCurationAllowed = useSelector(state => state.auth.user?.isCurationAllowed);
     return (
         <>
-            {sort === 'top' && (
-                <UncontrolledButtonDropdown>
-                    <DropdownToggle caret className="ps-3 pe-3" size="sm" color="light">
-                        {stringifySort(sort)}
-                    </DropdownToggle>
-                    <DropdownMenu>
-                        <DropdownItem disabled={isLoading} onClick={() => setSort('newest')}>
-                            Newest first
-                        </DropdownItem>
-                        <DropdownItem disabled={isLoading} onClick={() => setSort('oldest')}>
-                            Oldest first
-                        </DropdownItem>
-                    </DropdownMenu>
-                </UncontrolledButtonDropdown>
-            )}
-            {sort !== 'top' && (
-                <Tippy
-                    interactive={true}
-                    trigger="click"
-                    placement="bottom-end"
-                    onCreate={instance => setTippy(instance)}
-                    content={
-                        <div className="p-2">
-                            <FormGroup>
-                                <Label for="sortPapers">Sort</Label>
+            <Tippy
+                interactive={true}
+                trigger="click"
+                placement="bottom-end"
+                onCreate={instance => setTippy(instance)}
+                content={
+                    <div className="p-2">
+                        <FormGroup>
+                            <Label for="sortPapers">Sort</Label>
+                            <Input
+                                value={sort}
+                                onChange={e => {
+                                    tippy.hide();
+                                    setSort(e.target.value);
+                                }}
+                                bsSize="sm"
+                                type="select"
+                                name="sort"
+                                id="sortPapers"
+                                disabled={isLoading}
+                            >
+                                <option value="combined">Top recent</option>
+                                <option value="newest">Recently added</option>
+                                <option value="featured">Featured</option>
+                                {isCurationAllowed && <option value="unlisted">Unlisted</option>}
+                            </Input>
+                        </FormGroup>
+                        <FormGroup check>
+                            <Label check>
                                 <Input
-                                    value={sort}
                                     onChange={e => {
                                         tippy.hide();
-                                        setSort(e.target.value);
+                                        setIncludeSubFields(e.target.checked);
                                     }}
-                                    bsSize="sm"
-                                    type="select"
-                                    name="sort"
-                                    id="sortPapers"
+                                    checked={includeSubFields}
+                                    type="checkbox"
+                                    style={{ marginTop: '0.1rem' }}
                                     disabled={isLoading}
-                                >
-                                    <option value="newest">Newest first</option>
-                                    <option value="oldest">Oldest first</option>
-                                </Input>
-                            </FormGroup>
-                            <FormGroup check>
-                                <Label check>
-                                    <Input
-                                        onChange={e => {
-                                            tippy.hide();
-                                            setIncludeSubFields(e.target.checked);
-                                        }}
-                                        checked={includeSubFields}
-                                        type="checkbox"
-                                        style={{ marginTop: '0.1rem' }}
-                                        disabled={isLoading}
-                                    />
-                                    Include subfields
-                                </Label>
-                            </FormGroup>
-                        </div>
-                    }
-                >
-                    <span>
-                        <Button color="light" className="flex-shrink-0 ps-3 pe-3" style={{ marginLeft: 'auto' }} size="sm">
-                            {stringifySort(sort)} <Icon icon={faChevronDown} />
-                        </Button>
-                    </span>
-                </Tippy>
-            )}
+                                />
+                                Include subfields
+                            </Label>
+                        </FormGroup>
+                    </div>
+                }
+            >
+                <span>
+                    <Button color="light" className="flex-shrink-0 ps-3 pe-3" style={{ marginLeft: 'auto' }} size="sm">
+                        {stringifySort(sort)} <Icon icon={faChevronDown} />
+                    </Button>
+                </span>
+            </Tippy>
         </>
     );
 };