+++ src/pages/LiteratureList/LiteratureLists.js
@@ -3,15 +3,15 @@ import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import ListPage from 'components/ListPage/ListPage';
 import ListCard from 'components/LiteratureList/ListCard';
 import RequireAuthentication from 'components/RequireAuthentication/RequireAuthentication';
-import { CLASSES, PREDICATES } from 'constants/graphSettings';
+import { CLASSES } from 'constants/graphSettings';
 import ROUTES from 'constants/routes';
 import { groupBy } from 'lodash';
 import { reverse } from 'named-urls';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { getLiteratureListData } from 'utils';
 import { getResourcesByClass } from 'services/backend/resources';
 import { getStatementsBySubjects } from 'services/backend/statements';
-import { filterObjectOfStatementsByPredicateAndClass } from 'utils';
 
 const LiteratureLists = () => {
     const user = useSelector(state => state.auth.user);
@@ -32,26 +32,10 @@ const LiteratureLists = () => {
         if (resources.length) {
             items = await getStatementsBySubjects({ ids: resources.map(item => item.id) }).then(statements =>
                 statements.map(statementsForSubject => {
-                    const resource = resources.find(_resource => _resource.id === statementsForSubject.id);
-                    const _statements = statementsForSubject.statements;
-                    const description = filterObjectOfStatementsByPredicateAndClass(_statements, PREDICATES.DESCRIPTION, true);
-                    const listId = filterObjectOfStatementsByPredicateAndClass(_statements, PREDICATES.HAS_LIST, true)?.id;
-                    const researchField = filterObjectOfStatementsByPredicateAndClass(
-                        _statements,
-                        PREDICATES.HAS_RESEARCH_FIELD,
-                        true,
-                        CLASSES.RESEARCH_FIELD
+                    return getLiteratureListData(
+                        resources.find(_resource => _resource.id === statementsForSubject.id),
+                        statementsForSubject.statements
                     );
-                    const authors = filterObjectOfStatementsByPredicateAndClass(_statements, PREDICATES.HAS_AUTHOR, false);
-                    return {
-                        ...resource,
-                        id: resource.id,
-                        label: resource.label ? resource.label : 'No Title',
-                        description: description?.label ?? '',
-                        researchField,
-                        authors,
-                        listId
-                    };
                 })
             );
             const groupedByPaper = groupBy(items, 'listId');