+++ src/services/backend/papers.js
@@ -3,13 +3,16 @@ import { submitPutRequest, submitDeleteRequest, submitPostRequest, submitGetRequ
 import { getStatementsBySubjectAndPredicate } from 'services/backend/statements';
 import { PREDICATES } from 'constants/graphSettings';
 import { indexContribution } from 'services/similarity';
+import { toast } from 'react-toastify';
 
 export const papersUrl = `${url}papers/`;
 
 // Save full paper and index contributions in the similarity service
 export const saveFullPaper = (data, mergeIfExists = false) => {
-    return submitPostRequest(`${papersUrl}?mergeIfExists=${mergeIfExists}`, { 'Content-Type': 'application/json' }, data).then(paper => {
-        indexContributionsByPaperId(paper.id);
+    return submitPostRequest(`${papersUrl}?mergeIfExists=${mergeIfExists}`, { 'Content-Type': 'application/json' }, data).then(async paper => {
+        Promise.all(await indexContributionsByPaperId(paper.id)).catch(() =>
+            toast.warning('Similarity service seems to be down, skipping paper indexing')
+        );
         return paper;
     });
 };